requirements.txt
anyio==4.12.1
asgiref==3.11.1
beautifulsoup4==4.14.3
certifi==2026.1.4
charset-normalizer==3.4.4
Django==6.0.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
python-dotenv==1.2.1
python-telegram-bot==22.6
requests==2.32.5
soupsieve==2.8.3
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
whitenoise==6.9.0
psycopg2-binary==2.9.10
dj-database-url==2.3.0
telethon==1.36.0
